requirements.txt
# Core LLM / Agent framework
langchain-groq>=0.1.9
langchain-core>=0.2.0
langgraph>=0.2.0

# Browser automation
selenium>=4.15.0
webdriver-manager>=4.0.1

# DOM parsing
beautifulsoup4>=4.12.0

# Report generation
Jinja2>=3.1.0

# Config
python-dotenv>=1.0.0

# Utilities (already in stdlib but listed for clarity)
# logging, json, os, datetime, argparse, re — stdlib
